(PC-14439) feat(accessibility): add ButtonInsideText web
