+++ src/ui/components/buttons/buttonInsideText/ButtonInsideText.web.tsx
@@ -1,0 +1,44 @@
+import React from 'react'
+import styled, { CSSObject } from 'styled-components'
+
+import {
+  AppButtonEventNative,
+  TouchableOpacityButtonProps,
+} from 'ui/components/buttons/AppButton/types'
+import { ButtonInsideTextInner } from 'ui/components/buttons/buttonInsideText/ButtonInsideTextInner'
+import { ButtonInsideTexteProps } from 'ui/components/buttons/buttonInsideText/types'
+
+export function ButtonInsideText({
+  wording,
+  type,
+  onPress,
+  onLongPress,
+  icon: Icon,
+  color,
+  to,
+  externalHref = 'Hello',
+}: ButtonInsideTexteProps) {
+  const ButtonComponent = (to || externalHref ? Link : Button) as React.ElementType
+  return (
+    <ButtonComponent
+      onPress={onPress as AppButtonEventNative}
+      onLongPress={onLongPress as AppButtonEventNative}>
+      <ButtonInsideTextInner wording={wording} icon={Icon} color={color} type={type} />
+    </ButtonComponent>
+  )
+}
+
+const webStyle = {
+  display: 'inline-block',
+  border: 'none',
+  cursor: 'pointer',
+  outline: 'none',
+  textDecoration: 'none',
+  boxSizing: 'border-box',
+  backgroundColor: 'transparent',
+  margin: 0,
+  padding: 0,
+} as CSSObject
+
+const Button = styled.button<TouchableOpacityButtonProps>(webStyle)
+const Link = styled.a<TouchableOpacityButtonProps>(webStyle)
